fix: Combine spacing codemods and tests
